cryptfs: Fix mount failure when encryption triggered from settings
When the encryption is triggered from settings, the key is getting
created before the reboot and that key is not getting used to
encrypt the data. So create the key during the encryption not
before the encryption.

Change-Id: I9f4f244f433b16b5d319c9dbe757c8219a6cfc7f
(cherry picked from commit f14a3f2d47a28ad9ad0d999ad0ed52f6646ad333)